Warning fixes:         * Makefile.in (varasm.o): Depend on sdbout.h.         (sdbout.o): Depend on toplev.h.         * collect2.c (scan_prog_file): Cast fprintf argument to `long' and         use %ld specifier.         * final.c (shorten_branches): Cast first arg of `bzero' to char *.         * genextract.c (main): When creating insn-extract.c, mark variable         `i' with ATTRIBUTE_UNUSED.         * genpeep.c (main): When creating insn-peep.c, mark variables         `insn', `x' and `pat' with ATTRIBUTE_UNUSED.         * objc/init.c (__objc_tree_print): Wrap function definition in         macro `DEBUG'.         * objc/objc-act.c (encode_array): Cast sprintf argument to `long'         and use %ld specifier.         (adorn_decl): Likewise, twice.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@20650 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/collect2.c
@@ -2828,8 +2828,8 @@ scan_prog_file (prog_name, which_pass)
 #else
 		      if (debug)
 			fprintf (stderr,
-				 "\tiss = %5d, value = %5d, index = %5d, name = %s\n",
-				 symbol.iss, symbol.value, symbol.index, name);
+				 "\tiss = %5d, value = %5ld, index = %5d, name = %s\n",
+				 symbol.iss, (long) symbol.value, symbol.index, name);
 #endif
 		    }
 #ifdef COLLECT_EXPORT_LIST
